kingfisher: reboot: fix power-off on reset
Method cpld_reset_cpu of bl31 is called from the Linux kernel and uses
GPIO6, GPIO2 pins as SPI bus lines to control the CPLD device. But in the
kernel GPIO6_8 pin are initialized to work in interrupt mode instead of
the input/output mode. This leads to the fact that the SPI bus becomes
non-functional. In this patch we switch the GPIO6_8 pin back to the
input-output mode.

Jira-Id: ADM-773
Change-Id: Ic6af62398122585f3efd7e7257d32c87143537c5
Signed-off-by: Sergii Boryshchenko <sergii.boryshchenko@globallogic.com>