bl2: Pass Board ID to u-boot
u-boot needs to know the board ID
for loading proper dts file. This
has to be done before I2C initialization
in u-boot. Since board ID is stored
in eeprom which is accessible by I2C,
the only way to get the board ID
is to pass it from BL2.

Jira-Id: ADM-2684

Change-Id: I36978e2a5a0b6231bd62788c545f57156e839912
Signed-off-by: Vitaliy Vasylskyy <vitaliy.vasylskyy@globallogic.com>